+++ cli/src/pcluster/api/controllers/cluster_operations_controller.py
@@ -265,18 +265,27 @@ def describe_cluster(cluster_name, region=None):
 
 def _get_login_nodes(cluster):
     login_nodes_status = cluster.login_nodes_status
+
+    # TODO Fix once the API models are updated to support multiple pools in describe-cluster response
     if login_nodes_status.get_login_nodes_pool_available():
-        status = LoginNodesState.FAILED
-        if login_nodes_status.get_status() == LoginNodesPoolState.ACTIVE:
-            status = LoginNodesState.ACTIVE
-        elif login_nodes_status.get_status() == LoginNodesPoolState.PENDING:
-            status = LoginNodesState.PENDING
-        login_nodes = LoginNodesPool(status=status)
-        login_nodes.address = login_nodes_status.get_address()
-        login_nodes.scheme = login_nodes_status.get_scheme()
-        login_nodes.healthy_nodes = login_nodes_status.get_healthy_nodes()
-        login_nodes.unhealthy_nodes = login_nodes_status.get_unhealthy_nodes()
-        return login_nodes
+        login_nodes = []
+
+        for _pool_name, pool_status in login_nodes_status.get_pool_status_dict().items():
+            status = LoginNodesState.FAILED
+            if pool_status.get_status() == LoginNodesPoolState.ACTIVE:
+                status = LoginNodesState.ACTIVE
+            elif pool_status.get_status() == LoginNodesPoolState.PENDING:
+                status = LoginNodesState.PENDING
+            pool = LoginNodesPool(status=status)
+            # pool.name = pool_name
+            pool.address = pool_status.get_address()
+            pool.scheme = pool_status.get_scheme()
+            pool.healthy_nodes = pool_status.get_healthy_nodes()
+            pool.unhealthy_nodes = pool_status.get_unhealthy_nodes()
+            login_nodes.append(pool)
+            break
+
+        return login_nodes[0]
     return None