+++ cli/src/pcluster/config/cluster_config.py
@@ -1406,6 +1406,11 @@ class LoginNodes(Resource):
         super().__init__(**kwargs)
         self.pools = pools
 
+    def _register_validators(self, context: ValidatorContext = None):  # noqa: D102 #pylint: disable=unused-argument
+        self._register_validator(
+            DuplicateNameValidator, name_list=[pool.name for pool in self.pools], resource_name="Pool"
+        )
+
 
 class HeadNode(Resource):
     """Represent the Head Node resource."""