+++ cli/src/pcluster/models/cluster.py
@@ -284,12 +284,11 @@ class Cluster:
 
     @property
     def login_nodes_status(self):
-        """Status of the login nodes pool."""
+        """Status of the login nodes."""
         login_nodes_status = LoginNodesStatus(self.stack_name)
         if self.stack.scheduler == "slurm" and self.config.login_nodes:
-            # This approach works since by design we have now only one pool.
-            # We should fix this if we want to add more than a login nodes pool per cluster.
-            login_nodes_status.retrieve_data(self.config.login_nodes.pools[0].name)
+            login_node_pool_names = [pool.name for pool in self.config.login_nodes.pools]
+            login_nodes_status.retrieve_data(login_node_pool_names)
         return login_nodes_status
 
     @property
@@ -755,13 +754,18 @@ class Cluster:
                 )
         return self.__has_running_capacity
 
-    def has_running_login_nodes(self, updated_value: bool = False) -> bool:
-        """Return True if the cluster has running login nodes. Note: the value will be cached."""
+    def has_running_login_nodes(self, updated_value: bool = False, pool_name: str = None) -> bool:
+        """
+        Return True if the cluster has running login nodes, or a specific pool if a pool name is provided.
+
+        Note: the value will be cached.
+        """
+        healthy_nodes = self.login_nodes_status.get_healthy_nodes(pool_name=pool_name)
+        unhealthy_nodes = self.login_nodes_status.get_unhealthy_nodes(pool_name=pool_name)
+
         if self.__has_running_login_nodes is None or updated_value:
             self.__has_running_login_nodes = (
-                self.login_nodes_status.get_healthy_nodes() is not None
-                and self.login_nodes_status.get_unhealthy_nodes() is not None
-                and self.login_nodes_status.get_healthy_nodes() + self.login_nodes_status.get_unhealthy_nodes() != 0
+                healthy_nodes is not None and unhealthy_nodes is not None and healthy_nodes + unhealthy_nodes != 0
             )
         return self.__has_running_login_nodes